Moved all route definitions to api.py.
